Generate idmap files via proxy program, just like dexopt.
This commit changes how idmap files are generated by introducing a new
program, /system/bin/idmap, responsible for idmap creation. The tool
may be invoked similar to dexopt, ie via installd or as a stand-alone
program.

The stand-alone program allows creation of idmap files at any given
moment during runtime (eg package installation), rather than during boot
only.

Note: this commit depends on project platform/build to include the
addition of module `idmap'.

+++ include/utils/ResourceTypes.h
@@ -2067,8 +2067,9 @@ public:
     // Return value: on success: NO_ERROR; caller is responsible for free-ing
     // outData (using free(3)). On failure, any status_t value other than
     // NO_ERROR; the caller should not free outData.
-    status_t createIdmap(const ResTable& overlay, uint32_t originalCrc, uint32_t overlayCrc,
-                         void** outData, size_t* outSize) const;
+    status_t createIdmap(const ResTable& overlay,
+            uint32_t originalCrc, uint32_t overlayCrc,
+            void** outData, uint32_t* outSize) const;
 
     enum {
         IDMAP_HEADER_SIZE_BYTES = 3 * sizeof(uint32_t),
